Pulled out yet methods from extract_experiment_XML_vals.py to a new file SRA_EXPERIMENT_OBJ.py. Also added trivial method for pretty printing the combined platform and instrument data structure
